ppo_sac/requirements_spgym.txt
absl-py==1.4.0
appdirs==1.4.4
astunparse==1.6.3
box2d-py==2.3.5
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.4
commonmark==0.9.1
cycler==0.11.0
decorator==4.4.2
docker-pycreds==0.4.0
docstring-parser==0.15
etils==1.5.2
Farama-Notifications==0.0.4
ffmpeg==1.4
filelock==3.12.0
flatbuffers==24.3.25
fonttools==4.38.0
fsspec==2024.10.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
glfw==2.7.0
grpcio==1.54.0
gym==0.23.1
gym-dmc==0.3.0
gym-notices==0.0.8
gymnasium==1.0.0
h5py==3.12.1
huggingface-hub==0.11.1
idna==3.4
imageio==2.28.1
imageio-ffmpeg==0.3.0
importlib-metadata==5.2.0
importlib_resources==6.4.5
jax-jumpy==1.0.0
Jinja2==3.1.4
keras==2.11.0
kiwisolver==1.4.4
labmaze==1.0.6
libclang==18.1.1
lxml==5.3.0
Markdown==3.3.7
MarkupSafe==2.1.2
matplotlib==3.5.3
ml-dtypes==0.4.1
moviepy==1.0.3
mpmath==1.3.0
namex==0.0.8
networkx==3.2.1
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.9.0.80
opt_einsum==3.4.0
optree==0.13.1
orjson==3.10.11
packaging==23.1
pandas==2.2.3
pathtools==0.1.2
Pillow==9.5.0
proglog==0.1.10
protobuf==3.19.6
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pygame==2.6.1
Pygments==2.15.1
PyOpenGL==3.1.7
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.30.0
requests-oauthlib==1.3.1
rich==11.2.0
rsa==4.7.2
scipy==1.13.1
sentry-sdk==1.22.2
setproctitle==1.3.2
Shimmy==2.0.0
shtab==1.6.4
git+http://github.com/bryanoliveira/sliding-puzzle-env.git
smmap==5.0.0
stable_baselines3==2.4.0
swig==4.3.0
sympy==1.13.1
tenacity==8.2.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensordict==0.6.2
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
torchrl==0.6.0
tqdm==4.65.0
triton==3.1.0
typing_extensions==4.12.2
tyro==0.5.10
tzdata==2024.2
urllib3==1.26.15
wandb==0.13.11
Werkzeug==2.2.3
wrapt==1.17.0
zipp==3.15.0